requirements.txt
streamlit
google-genai
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
python-dotenv